% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ramka.R
\name{generuj_ramke_d1_parametry}
\alias{generuj_ramke_d1_parametry}
\title{Generowanie parametrów dla wskaźnika D1}
\usage{
generuj_ramke_d1_parametry(glowna_edycja)
}
\arguments{
\item{glowna_edycja}{Liczba całkowita reprezentująca główną edycję (rok),
dla której generowane są parametry. Np. 2025.}
}
\value{
Ramka danych typu \code{tibble} zawierająca kolumny:
\code{wskaznik} (z wartością "D1"), \code{WOJ_NAZWA}, \code{typ_szk2}, \code{rok_abs},
\code{edycja}, \code{kryterium} oraz \code{wynik} (inicjowany jako \code{NULL}).
}
\description{
Funkcja tworzy ramkę danych zawierającą unikalne kombinacje parametrów
dla wskaźnika D1. Parametry te posłużą do późniejszego generowania rozkładów.
Wskaźnik D1 jest obliczany tylko dla typów szkół \strong{niebędących}
"Liceum ogólnokształcące", "Liceum dla dorosłych" i "Szkoła specjalna przysposabiająca do pracy".
Rok akademicki (\code{rok_abs}) jest obliczany na podstawie głównej edycji
(parametr \code{glowna_edycja}), przyjmując wartości \code{edycja - 1}, \code{edycja - 2} i \code{edycja - 5}.
Dla tych typów szkół zmienna \code{kryterium} zawsze przyjmuje wartości "sexf" i "nazwa_zaw".
}
