% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ramka.R
\name{generuj_ramke_k1_parametry}
\alias{generuj_ramke_k1_parametry}
\title{Generowanie parametrów dla wskaźnika K1}
\usage{
generuj_ramke_k1_parametry(glowna_edycja)
}
\arguments{
\item{glowna_edycja}{Liczba całkowita reprezentująca główną edycję (rok),
dla której generowane są parametry. Np. 2025.}
}
\value{
Ramka danych typu \code{tibble} zawierająca kolumny:
\code{wskaznik} (z wartością "K1"), \code{WOJ_NAZWA}, \code{typ_szk2}, \code{rok_abs},
\code{edycja}, \code{kryterium}, \code{typ_k1_analizy} (identyfikator kategorii analizy K1),
\code{parametr_k1_wartosci} (lista wartości specyficznych dla kategorii K1)
oraz \code{wynik} (inicjowany jako \code{NULL}).
}
\description{
Funkcja tworzy ramkę danych zawierającą unikalne kombinacje parametrów
dla wskaźnika K1. Wskaźnik K1 posiada dwie kategorie analizy
("kategoria_k1_1" i "kategoria_k1_2"), z których każda ma własny zestaw
wartości do wykorzystania w analizie (przechowywany w kolumnie \code{parametr_k1_wartosci}).
Rok ukończenia szkoły (\code{rok_abs}) jest obliczany na podstawie głównej edycji
(parametr \code{glowna_edycja}), przyjmując wartości \code{edycja - 1}, \code{edycja - 2} i \code{edycja - 5}.
Typy szkół obejmują wszystkie dostępne typy z listy \code{TYPY_SZKOL}.
Zmienna \code{kryterium} jest dobierana dynamicznie w zależności od typu szkoły.
}
