% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ramka.R
\name{generuj_ramke_migracje}
\alias{generuj_ramke_migracje}
\title{Generowanie parametrów ogólnych dla wskaźników z typ_szk2 = NULL
i kryterium = teryt_woj dla wybranej grupy szkół}
\usage{
generuj_ramke_migracje(glowna_edycja)
}
\arguments{
\item{glowna_edycja}{Liczba całkowita reprezentująca główną edycję (rok),
dla której generowane są parametry. Np. 2025.}
}
\value{
Ramka danych typu \code{tibble} zawierająca kolumny:
\code{wskaznik} (z wartością "meta_zaw"), \code{WOJ_NAZWA}, \code{typ_szk2} (NULL), \code{rok_abs},
\code{edycja}, \code{kryterium} (z wartością "teryt_woj"), \code{rok}, \code{miesaic}
oraz \code{wynik} (inicjowany jako \code{NULL}).
}
\description{
Funkcja tworzy ramkę danych dla ogólnego wskaźnika obejmującego
wszystkie szkoły, gdzie typ_szk2 jest traktowany jako NULL
(agregacja na poziomie województwa), a kryterium to
"teryt_woj". Jest to użyteczne dla analiz migracji.
}
