% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Przygotowanie_tabel_pośrednich_pod_pakiet.R
\name{przetwarzaj_p2}
\alias{przetwarzaj_p2}
\title{Przekształcenie tabeli p2 do formatu szerokiego
#' @description Ta funkcja przetwarza ramkę danych 'p2' w celu zsumowania
wystąpień dziedzin i dyscyplin dla każdego absolwenta, roku i okresu.}
\usage{
przetwarzaj_p2(p2)
}
\arguments{
\item{p2}{Ramka danych 'p2' do przetworzenia.}
}
\value{
Zwraca przetworzoną ramkę danych.
}
\description{
Przekształcenie tabeli p2 do formatu szerokiego
#' @description Ta funkcja przetwarza ramkę danych 'p2' w celu zsumowania
wystąpień dziedzin i dyscyplin dla każdego absolwenta, roku i okresu.
}
