% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Przygotowanie_tabel_pośrednich_pod_pakiet.R
\name{przygotuj_p2p3p4}
\alias{przygotuj_p2p3p4}
\title{Przygotowanie zbiorów danych p2, p3 i p4}
\usage{
przygotuj_p2p3p4(p2, p3, p4)
}
\arguments{
\item{p2}{Ramka danych w formacie tibble.}

\item{p3}{Ramka danych w formacie tibble.}

\item{p4}{Ramka danych w formacie tibble.}
}
\value{
Zwraca listę z trzema ramkami danych: p2_1_light, p3_light, p4_light.
}
\description{
Funkcja, która przetwarza ramki danych p2, p3 i p4,
tworząc z nich lżejsze wersje z wybranymi i przekształconymi
kolumnami.
}
