% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_freq}
\alias{tab_freq}
\title{funkcja tworzaca tabele z odsetkiem w kategoriach zmiennej (pierwotnie
do tworzenia tabel do migracji)}
\usage{
tab_freq(dat, kryterium)
}
\arguments{
\item{dat}{dane na podstawie ktorych ma powstac tabela}

\item{kryterium}{zmienna której częstości mają być wyliczone}
}
\description{
funkcja tworzaca tabele z odsetkiem w kategoriach zmiennej (pierwotnie
do tworzenia tabel do migracji)
}
