% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_kolu_K1}
\alias{tab_kolu_K1}
\title{Tabela z kontynuowaniem nauki w różnych formach, ktore musza byc
przekazane w liscie wskaznikow, ulozonych w kolumnach w podziale na
kryterium w wierszach (pierwotnie przygotowane pod zawody)}
\usage{
tab_kolu_K1(dat, wskazniki, kryterium)
}
\arguments{
\item{dat}{dane na podstawie ktorych ma powstac tabela - dane te są
przekazywane do funkcji tab_wier}

\item{wskazniki}{lista wskaznikow, z których ma powstać tabela wynikowa ze
sposobami kontynuowania nauki - kazdy wskaznik jest osobno przekazywany do
funkcji tab_wier}

\item{kryterium}{zmienna przez która ma być skrosowany wskaznik - zmienna ta
jest przekazywana do funkcji tab_wier}
}
\description{
Zastosowanie funkcji  pozwala uzyskac tabele,
w ktorej przechowywana jest informacja o liczbie i procencie kobiet, mezczyzn
i ogolem wszystkich absolwentow kontynuujacych naukę na studiach, w szkole
policealnej lub na KKZ
}
