% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_msc}
\alias{tab_msc}
\title{funkcja tworzaca tabele krzyzowa, w ktorej wskaznik jest w wierszach a
w kolumnach jest zmienna czasu "mscrok"}
\usage{
tab_msc(dat, wskaznik, kryterium, edycja)
}
\arguments{
\item{dat}{dane na podstawie ktorych ma powstac tabela}

\item{wskaznik}{wskaznik, z którego ma powstać tabela}

\item{kryterium}{zmienna przez która ma być skrosowany wskaznik}

\item{edycja}{parametr przekazujący dla której edycji monitoringu ma być
tworzona tabela; druga zmienna, ktora umożliwia utworzenie tabeli "rok_abs"
w przekazanych danych moze posiadac tylko jedną wartosc - musi byc
odfiltrowana wczesniej}
}
\description{
funkcja tworzaca tabele krzyzowa, w ktorej wskaznik jest w wierszach a
w kolumnach jest zmienna czasu "mscrok"
}
