% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_sz_sex}
\alias{tab_sz_sex}
\title{funkcja tworzaca tabele krzyzowa, typ szkoły przez płeć z procentami
po ogółem typie szkoły}
\usage{
tab_sz_sex(dat)
}
\arguments{
\item{dat}{dane na podstawie ktorych ma powstac tabela domyślnie p4 lub jakiś
fragment p4 po odfiltrowaniu}
}
\description{
funkcja tworzaca tabele krzyzowa, typ szkoły przez płeć z procentami
po ogółem typie szkoły
}
