% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_sz_zaw}
\alias{tab_sz_zaw}
\title{funkcja tworzaca tabele z udziałem 10 najczęstszych zawodów w szkołąch
kształcących zawodowo z podziałem na młodocianych i niemłodocianych}
\usage{
tab_sz_zaw(dat)
}
\arguments{
\item{dat}{dane na podstawie ktorych ma powstac tabela powinno być p4 całe
lub po filtrze}
}
\description{
funkcja tworzaca tabele z udziałem 10 najczęstszych zawodów w szkołąch
kształcących zawodowo z podziałem na młodocianych i niemłodocianych
}
