% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_wier}
\alias{tab_wier}
\title{funkcja tworzaca tabele krzyzowa, w ktorej wskaznik jest w wierszach}
\usage{
tab_wier(dat, wskaznik, kryterium)
}
\arguments{
\item{dat}{dane na podstawie ktorych ma powstac tabela}

\item{wskaznik}{wskaznik, z którego ma powstać tabela}

\item{kryterium}{zmienna przez która ma być skrosowany wskaznik}
}
\description{
funkcja tworzaca tabele krzyzowa, w ktorej wskaznik jest w wierszach
}
