% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_wier_K2}
\alias{tab_wier_K2}
\title{Tabela z kontynuowaniem nauki w różnych dziedzinach lub dyscyplinach,
ulozonych w kolumnach w podziale na kryterium
(z zalozenia plec lub nazwa_zaw) w wierszach}
\usage{
tab_wier_K2(dat, wskaznik, kryterium)
}
\arguments{
\item{dat}{dane bedące odfiltrowanym w funkcjach nadrzednych fragmentem
tabeli p2_1 przygotowanej na potrzeby raportow z tabeli p2}

\item{wskaznik}{wskaznik może przybierać tylko dwie wartości dziedziny lub
dyscypliny}

\item{kryterium}{zmienna przez która ma być skrosowany wskaznik -
może przybierać dwie wartości plec lub nazwa_zaw}
}
\description{
Zastosowanie funkcji  pozwala uzyskac tabele,
w ktorej przechowywana jest informacja o liczbie i procencie ze względu
na płeć lub wyuczony zawód oraz i ogółem wszystkich absolwentóww
kontynuujacych naukę na studiach w różnych dziedzinach/dyscyplinach
}
