% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_sex.R
\name{tab_wyna}
\alias{tab_wyna}
\title{funkcja tworzaca tabele ze statystykami wynagrodzen w poidziale na
kryterium (płeć lub zawód)}
\usage{
tab_wyna(dat, kryterium)
}
\arguments{
\item{dat}{dane na podstawie ktorych ma powstac tabela bedące podzbiorem
tabeli p3wyodrębnionym ze względu na typ szkoły, (ewentualnie) województwo,
rok absolwenta oraz okres za który mają być uśrednione zarobki}

\item{kryterium}{zmienna przez która ma być skrosowany wskaznik domyslnie
płeć lub nazwa zawodu}
}
\description{
funkcja tworzaca tabele ze statystykami wynagrodzen w poidziale na
kryterium (płeć lub zawód)
}
