% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wynik_tab_paralel.R
\name{wypelnij_wyniki_branzowe_rownolegle}
\alias{wypelnij_wyniki_branzowe_rownolegle}
\title{Wypełnia kolumnę 'wynik' tabeli parametrów wygenerowanymi tabelami krzyżowymi}
\usage{
wypelnij_wyniki_branzowe_rownolegle(param_df, p2_1, p3, p4)
}
\arguments{
\item{param_df}{Ramka danych z parametrami wskaźników (np. wygenerowana przez \code{polacz_ramki_wskaznikow}).
Powinna zawierać kolumny: \code{wskaznik}, \code{branza}, \code{typ_szk2}, \code{rok_abs}, \code{edycja}, \code{kryterium},
\code{rok}, \code{miesiac}, \code{typ_k1_analizy}, \code{parametr_K1}, \code{parametr_K2}, \code{parametr_grupa_szkol}.}

\item{p2_1}{Dane źródłowe dla wskaźnika K2.}

\item{p3}{Dane źródłowe dla wskaźników S7, K1, W1, B1.}

\item{p4}{Dane źródłowe dla wskaźników D1, D2, typ_szk2, meta_zaw.}
}
\value{
Zmodyfikowana ramka danych \code{param_df} z wypełnioną kolumną \code{wynik},
gdzie każda komórka zawiera tabelę krzyżową lub informację o błędzie/braku danych.
}
\description{
Funkcja iteruje przez wiersze tabeli parametrów i dla każdej kombinacji
parametru (wskaźnik, kryterium, filtry, itd.) filtruje odpowiednie dane źródłowe
i wywołuje właściwą funkcję generującą tabelę krzyżową. Wynik tabeli krzyżowej
jest zapisywany w kolumnie 'wynik' jako element listy.
}
